+++ src/Conversion/ONNXToTorch/CMakeLists.txt
@@ -16,7 +16,7 @@ add_onnx_mlir_library(OMONNXToTorch
   NN/AddOp.cpp
   NN/FlattenOp.cpp
   NN/ElementwiseOp.cpp
-  NN/Identity.cpp
+  NN/IdentityOp.cpp
 
   LINK_LIBS PUBLIC
   OMConstPropHelper
